fix "overloading" in production build
As pointed out in https://github.com/w3c/ServiceWorker/issues/339
there is no way (at the current date) to bypass the automatic
unzipping when receiving a gzip content, which generates the issue
of the file being passed to the client with the size of the compressed
data and the bytes in the pregression were being counted after
uncompressing them, leading to the content length mismatch and the
loading percentage going over 100.
One option to bypass this issue would also be setting the
Accept-Encoding Header to the identity and issue a head request to only
get the actual size of the data, but as pointed out in
https://fetch.spec.whatwg.org/#terminology-headers only the user agent
has access and full control of this header.
The current solution is only a workaround and should be solved in the
backend instead of statically putting the original file size in the
nginx config.